feat: :sparkles: Finish Header.
Finish Social. Finish Header. Update print stylings.
