+++ src/components/atoms/profilePicture/ProfilePicture.module.css
@@ -1,0 +1,11 @@
+.profilePicture {
+	/* Match the height of the title */
+	height: 6rem;
+
+	/* Subtle rounding of the corners to make it less perfect. */
+	border-radius: 2px;
+
+	/* Black & White */
+	-webkit-filter: grayscale(100%);
+	filter: grayscale(100%);
+}
