+++ src/components/atoms/title/Title.module.css
@@ -1,0 +1,21 @@
+.title {
+	/* Font sizing */
+	font-weight: 900;
+	font-size: 3rem;
+	text-transform: uppercase;
+
+	/* Keeps the spacing between lines much smaller than default. */
+	line-height: 3rem;
+	margin: 0;
+
+	/* Sets the with to the shortest word in the title. Breaks line after each word. */
+	white-space: pre-line;
+	width: min-content;
+}
+
+/* Makes the title a little smaller for the print version. */
+@media print {
+	.title {
+		font-size: 2rem;
+	}
+}
