+++ src/components/index.ts
@@ -1,5 +1,5 @@
-export * from './base/Base';
-export * from './header/Header';
-export * from './profilePicture/ProfilePicture';
-export * from './title/Title';
-export * from './social/Social';
+export * from './atoms/base/Base';
+export * from './molecules/header/Header';
+export * from './atoms/profilePicture/ProfilePicture';
+export * from './atoms/title/Title';
+export * from './atoms/social/Social';
