+++ src/components/molecules/header/Header.tsx
@@ -2,7 +2,12 @@ import { type FC } from 'react';
 // ! Must use the import from index here!
 // The problem is with vite-plugin-ssr which cannot handle import from "react-icons/bs".
 // It throws an error that it cannot import directory for some reason.
-import { BsGithub, BsLinkedin, BsTwitter } from 'react-icons/bs/index';
+import {
+	BsEnvelopeFill,
+	BsGithub,
+	BsLinkedin,
+	BsTwitter,
+} from 'react-icons/bs/index';
 
 import { ProfilePicture, Social, Title } from '@components';
 
@@ -13,25 +18,32 @@ export const Header: FC = () => {
 		<div className={style.header}>
 			<Title>Patrik Dvořáček</Title>
 
-			<div>
+			<div className={style.socials}>
+				<Social
+					title="Email"
+					link="mailto:patrik@dvoracek.dev"
+					username="patrik@dvoracek.dev"
+					icon={<BsEnvelopeFill />}
+				/>
+
 				<Social
 					title="GitHub"
 					link="https://github.com/tehSIRius"
-					username="tehSIRius"
+					username="@tehSIRius"
 					icon={<BsGithub />}
 				/>
 
 				<Social
 					title="Twitter"
 					link="https://twitter.com/DvoracekPatrik"
-					username="DvoracekPatrik"
+					username="@DvoracekPatrik"
 					icon={<BsTwitter />}
 				/>
 
 				<Social
 					title="LinkedIn"
 					link="https://www.linkedin.com/in/patrikdvoracek/"
-					username="patrikdvoracek"
+					username="@patrikdvoracek"
 					icon={<BsLinkedin />}
 				/>
 			</div>
