--- src/components/social/Social.module.css
@@ -1,4 +1,0 @@
-.social a {
-	text-decoration: none;
-	color: inherit;
-}
